FEAT: push slides and readme updated
