+++ .gitignore
@@ -7,6 +7,7 @@ __pycache__/
 *.so
 
 .idea/
+.DS_Store
 
 # Distribution / packaging
 .Python
