+++ groupe7_planification_edt/gui/main.py
@@ -1,0 +1,9 @@
+import sys
+from PyQt5.QtWidgets import QApplication
+from main_window import MainWindow
+
+if __name__ == '__main__':
+    app = QApplication(sys.argv)
+    window = MainWindow()
+    window.show()
+    sys.exit(app.exec_())
