--- groupe7_planification_edt/src/main.py
@@ -1,9 +1,0 @@
-import sys
-from PyQt5.QtWidgets import QApplication
-from main_window import MainWindow
-
-if __name__ == '__main__':
-    app = QApplication(sys.argv)
-    window = MainWindow()
-    window.show()
-    sys.exit(app.exec_())
