--- groupe7_planification_edt/src/styler.qss
@@ -1,68 +1,0 @@
-QMainWindow {
-    background-color: #f5f5f5;
-}
-
-QGroupBox {
-    background-color: #ffffff;
-    border: 1px solid #3498db;
-    border-radius: 5px;
-    margin-top: 10px;
-    padding: 10px;
-    color: #000000; /* Text inside group boxes set to black */
-}
-
-QGroupBox::title {
-    subcontrol-origin: margin;
-    left: 10px;
-    padding: 0 5px 0 5px;
-    color: #000000; /* Title text set to black */
-    font-size: 16px;
-    font-weight: bold;
-}
-
-QPushButton {
-    background-color: #3498db;
-    color: #fff;  /* Button text remains white for contrast on blue */
-    border: none;
-    border-radius: 5px;
-    padding: 8px 16px;
-    font-size: 14px;
-}
-
-QPushButton:hover {
-    background-color: #2980b9;
-}
-
-QTableWidget {
-    background-color: #ecf0f1;
-    border: 1px solid #bdc3c7;
-    font-size: 14px;
-    color: #000000; /* Table text set to black */
-}
-
-QHeaderView::section {
-    background-color: #3498db;
-    color: white;  /* Header remains white on blue */
-    padding: 4px;
-    border: none;
-    font-size: 14px;
-}
-
-QLabel {
-    font-size: 14px;
-    color: #000000; /* Label text set to black */
-}
-
-QTabBar::tab {
-    background-color: #ecf0f1; /* Or your preferred background */
-    color: #000000; /* Set tab text to black */
-    padding: 8px;
-    border: 1px solid #bdc3c7;
-    border-bottom: none;
-    font-size: 12px;
-}
-
-QTabBar::tab:selected {
-    background-color: #ffffff;
-    color: #000000;
-}
